requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.10       # PostgreSQL adapter
Flask-Login==0.6.3
Flask-Limiter[redis]==3.5.0   # Using Redis for robust rate limiting storage
Flask-Bcrypt==1.0.1           # For strong password hashing
Flask-Migrate==4.0.5          # Database migrations
Flask-WTF==1.2.1              # CSRF protection
Flask-CORS==4.0.0             # Cross-Origin Resource Sharing
Flask-Mail==0.10.0            # Email sending
python-dotenv==1.0.0
requests==2.31.0              # HTTP library for M-Pesa API calls
gunicorn==20.1.0
# Allow any compatible Celery 5.x release; some specific micro versions may be yanked
# or unavailable in certain installers. Using a permissive range improves deploy
# reliability while staying on Celery 5.
celery>=5.4,<6.0
redis==4.5.5
flask-caching==2.0.2
# Align OpenTelemetry API/SDK with the instrumentation 0.41b0 series
opentelemetry-instrumentation-flask==0.41b0
sentry-sdk==2.48.0
# Use a PyPI-available prometheus client
prometheus-client==0.21.1
prometheus_flask_exporter==0.23.2

phonenumbers>=8.12.0,<9.0

# Image processing for thumbnails
Pillow>=10.0.0
boto3>=1.28.0
